fix static-analysis violations in e-billing ship-to adapters and party address formatting
